+++ lib/index.js
@@ -24,14 +24,18 @@ const finished = [];
 async function fetchUUID(name) {
     const response = await fetch(`https://playerdb.co/api/player/minecraft/${name}`);
     const UUID = await response.json();
-    // console.log(await UUID.data.player.raw_id)
-    return await UUID.data.player.raw_id;
+    // console.log(await UUID.data.player.id)
+    return await UUID.data.player.id;
 }
 async function jsonGen() {
     for (const name of whitelist) {
         finished.push({
             uuid: `${await fetchUUID(name)}`,
             name: `${name}`,
+        });
+        finished.push({
+            uuid: `${await fetchUUID(name)}`,
+            name: `.${name}`,
         });
     }
     // console.log(finished);
