+++ src/whitelist.ts
@@ -1,0 +1,17 @@
+interface IWhitelist {
+    uuid: string;
+    name: string;
+  }
+  
+  const whitelist: string[] = [
+    "Cyb3rBac0n",
+    "Herobrine3953",
+    "DREW_BONKERS_",
+    "starpilot999",
+    "SoliasDream",
+    "FIREPILOT22",
+    "Flintsteel351",
+    "cytron499",
+  ];
+
+export {IWhitelist, whitelist}
